+++ lib/features/logistics/shipment/domain/params/create_shipment.dart
@@ -1,15 +1,11 @@
-class Shipment {
-  final String trackingId;
+class CreateShipmentRequest {
   final String customer;
   final String phone;
   final String address;
-  final String status;
 
-  const Shipment({
-    required this.trackingId,
+  const CreateShipmentRequest({
     required this.customer,
     required this.phone,
     required this.address,
-    required this.status,
   });
 }