+++ lib/features/logistics/shipment/domain/usecases/create_shipment_usecase.dart
@@ -1,0 +1,13 @@
+import 'package:logistic_operation/features/logistics/shipment/domain/params/create_shipment.dart';
+
+import '../repository/shipment_repository.dart';
+
+class CreateShipmentUseCase {
+  final ShipmentRepository repository;
+
+  const CreateShipmentUseCase(this.repository);
+
+  Future<void> call(CreateShipmentRequest request) {
+    return repository.createShipment(request);
+  }
+}