+++ lib/shared/widgets/app_text_field.dart
@@ -1,0 +1,41 @@
+import 'package:flutter/material.dart';
+
+class AppTextField extends StatelessWidget {
+  final TextEditingController controller;
+  final String label;
+  final String? hint;
+  final TextInputType keyboardType;
+  final String? Function(String?)? validator;
+  final int maxLines;
+  final bool obscureText;
+  final Widget? prefixIcon;
+
+  const AppTextField({
+    super.key,
+    required this.controller,
+    required this.label,
+    this.hint,
+    this.keyboardType = TextInputType.text,
+    this.validator,
+    this.maxLines = 1,
+    this.obscureText = false,
+    this.prefixIcon,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return TextFormField(
+      controller: controller,
+      keyboardType: keyboardType,
+      validator: validator,
+      obscureText: obscureText,
+      maxLines: maxLines,
+      decoration: InputDecoration(
+        labelText: label,
+        hintText: hint,
+        prefixIcon: prefixIcon,
+        border: OutlineInputBorder(borderRadius: BorderRadius.circular(12)),
+      ),
+    );
+  }
+}